Disable localhost being rewritten as more permissive URLACL
When "localhost" is specified the default nancy behaviour is to set the
UrlAcl to "http://+:1234/" rather than "http://localhost:1234".  The "+"
is more permissive and allows the Url to be accessed off the corrent
server.  By disabling this the computer name, "+" or "*" must be
explicitly used in the config.  This is in line with how SC works